fix(sandbox): strip ambient gateway token
